Get most (except DDA) of the instructions working correctly based on the sm83 tests

+++ src/registers.h
@@ -95,7 +95,7 @@ struct Registers {
     return result;
   }
 
-  inline uint8_t pop16(const uint8_t *mem) {
+  inline uint16_t pop16(const uint8_t *mem) {
     auto result = Mem::get16(mem, sp);
     sp += 2;
     return result;
